refactor(webapp): reorganize budget-details directory layout
