+++ frontend/projects/webapp/src/app/feature/budget/budget-details/allocated-transactions/create-dialog/dialog.ts
@@ -26,7 +26,7 @@ import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { TranslocoPipe } from '@jsverse/transloco';
 import { type BudgetLine, type TransactionCreate } from 'pulpe-shared';
-import { transactionCreateFromFormSchema } from '../../edit-transaction-form';
+import { transactionCreateFromFormSchema } from '../../components/edit-transaction-form';
 import { formatLocalDate } from '@core/date/format-local-date';
 import {
   applyAmountValidators,
