+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/budget-grid/budget-grid-mobile-card.ts
@@ -17,10 +17,10 @@ import { AppCurrencyPipe, FormatConversionPipe } from '@core/currency';
 import { FinancialKindDirective } from '@ui/financial-kind';
 import { FinancialLineCard } from '@pattern/financial-line-card';
 import { OriginalAmountLine } from '@ui/original-amount-line';
-import { formatMatchAnnotation } from '../view-models/budget-item-constants';
-import type { BudgetLineTableItem } from '../view-models/table-items.view-model';
-import { SegmentedBudgetProgress } from '../components/segmented-budget-progress';
-import { BudgetActionMenu } from '../components/budget-action-menu';
+import { formatMatchAnnotation } from '../../view-models/budget-item-constants';
+import type { BudgetLineTableItem } from '../../view-models/table-items.view-model';
+import { SegmentedBudgetProgress } from '../segmented-budget-progress';
+import { BudgetActionMenu } from '../budget-action-menu';
 
 @Component({
   selector: 'pulpe-budget-grid-mobile-card',
