+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/budget-grid/budget-grid.ts
@@ -21,9 +21,9 @@ import { FinancialKindIndicator } from '@ui/financial-kind-indicator';
 import { OriginalAmountLine } from '@ui/original-amount-line';
 import { TransactionLabelPipe } from '@ui/transaction-display';
 import type { BudgetLine, Transaction, SupportedCurrency } from 'pulpe-shared';
-import type { TransactionViewModel } from '../view-models/transaction.view-model';
-import type { BudgetLineTableItem } from '../view-models/table-items.view-model';
-import { TransactionActionMenu } from '../components/transaction-action-menu';
+import type { TransactionViewModel } from '../../view-models/transaction.view-model';
+import type { BudgetLineTableItem } from '../../view-models/table-items.view-model';
+import { TransactionActionMenu } from '../transaction-action-menu';
 import {
   BudgetDetailPanel,
   type BudgetDetailPanelData,
