+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/budget-table/budget-table-checked-filter.ts
