+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/budget-table/cells/actions-cell.ts
@@ -16,7 +16,7 @@ import type { BudgetLine } from 'pulpe-shared';
 import type {
   BudgetLineTableItem,
   TransactionTableItem,
-} from '../../view-models/table-items.view-model';
+} from '../../../view-models/table-items.view-model';
 
 @Component({
   selector: 'pulpe-actions-cell',
