+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/budget-table/cells/balance-cell.ts
@@ -11,7 +11,7 @@ import { UserSettingsStore } from '@core/user-settings';
 import type {
   BudgetLineTableItem,
   TransactionTableItem,
-} from '../../view-models/table-items.view-model';
+} from '../../../view-models/table-items.view-model';
 
 @Component({
   selector: 'pulpe-balance-cell',
