+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/index.ts
@@ -2,3 +2,4 @@ export { SegmentedBudgetProgress } from './segmented-budget-progress';
 export { BudgetActionMenu } from './budget-action-menu';
 export { BudgetViewToggle } from './budget-view-toggle';
 export { TransactionActionMenu } from './transaction-action-menu';
+export { BudgetItemsContainer } from './budget-items-container';
