+++ frontend/projects/webapp/src/app/feature/budget/budget-details/utils/budget-details-snackbar.utils.ts
